+++ billing/backend/internal/dto/ping.go
@@ -5,7 +5,7 @@ type PingResponse struct {
 }
 
 // NewPingResponse creates a new instance of PingResponse with the provided HTTP code, status, and message.
-func NewPingResponse(httpCode int16, status, message string) PingResponse {
+func NewPingResponse(httpCode int, status, message string) PingResponse {
 	return PingResponse{
 		ResponseBase: NewResponseBase(httpCode, status, message),
 	}
